Arquitetura e Design de Projetos Java/08_clean_architecture/escola/src/main/java/br/com/alura/escola/aplicacao/aluno/matricula/MatricularAluno.java
package br.com.alura.escola.aplicacao.aluno.matricula;

import br.com.alura.escola.dominio.aluno.Aluno;
import br.com.alura.escola.dominio.aluno.RepositorioDeAlunos;

public class MatricularAluno {
	// Use case
	// Padrão command
	private final RepositorioDeAlunos repositorio;

	public MatricularAluno(RepositorioDeAlunos repositorio) {
		this.repositorio = repositorio;
	}
	
	public void executar(MatricularAlunoDto dados) {
		Aluno novo = dados.criarAluno();
		
		this.repositorio.matricular(novo);
	}
}
